downsampling: Keep the intermediate scratch buffer in `f32`
The horizontal pass of the separable Lanczos filter wrote its result
back to `u8`, clamping and quantizing it before the vertical pass
consumed it.  Besides losing precision this clamps away the negative
lobes of the filter before they get a chance to cancel out in the
second pass.

Store the intermediate as unclamped `f32` instead.  It is a
prerequisite for doing the filtering in linear space for sRGB inputs,
where 8 bits of linear precision are wholly insufficient.  The 4x wider
scratch buffer costs 7-13% for `Unorm` and `Snorm` inputs:

    square_test.png           64.8ms -> 70.8ms  (+9.2%)
    2048->1024 rgb8_unorm     82.7ms -> 90.8ms  (+9.9%)
    2048->1024 rgba8_unorm   102.1ms -> 114.5ms (+12.2%)
    mip chain 2048 rgba8     137.2ms -> 155.6ms (+13.4%)

Alpha weighting used to resolve the alpha-weighted color after each
pass, which relied on that clamp: with an unclamped intermediate the
negative lobes can leave a horizontal alpha sum at or below zero,
selecting the unweighted fallback color, which then leaks into the
vertical pass with a negative weight.  Instead, keep the unresolved
alpha-weighted color sum, the alpha sum and the unweighted color sum
in the scratch buffer (7 floats per texel), accumulate them linearly in
the vertical pass and only resolve once at the end.  That costs ~30%
for alpha-weighted downsampling:

    2048->1024 rgba8_unorm_alpha_weighted  113.1ms -> 148.1ms (+31.0%)
    mip chain 2048 rgba8_alpha_weighted    152.8ms -> 198.1ms (+29.6%)

All measured with the prebuilt kernels from the generate-binaries
workflow.

+++ src/ispc/downsample_ispc.rs
@@ -252,7 +252,7 @@ fn bindgen_test_layout_SampleWeights() {
 #[derive(Debug, Copy, Clone)]
 pub struct DownsamplingContext {
     pub weights: SampleWeights,
-    pub scratch_space: *mut u8,
+    pub scratch_space: *mut f32,
 }
 #[test]
 fn bindgen_test_layout_DownsamplingContext() {
